requirements.txt
PyYAML==6.0.2
boto3==1.38.30
rich==14.0.0